FEATURE,PART3/3: Cleanup params and remove counts from v2.
