+++ services/indexes/params/params.go
@@ -9,6 +9,7 @@ import (
 	"net/url"
 	"time"
 
+	"github.com/ava-labs/avalanchego/ids"
 	"github.com/gocraft/dbr/v2"
 )
 
@@ -16,6 +17,7 @@ const (
 	KeyID              = "id"
 	KeyChainID         = "chainID"
 	KeyAddress         = "address"
+	KeyAlias           = "alias"
 	KeyAssetID         = "assetID"
 	KeySearchQuery     = "query"
 	KeySortBy          = "sort"
@@ -33,8 +35,6 @@ const (
 	PaginationMaxLimit      = 500
 	PaginationDefaultLimit  = 500
 	PaginationDefaultOffset = 0
-
-	MinSearchQueryLength = 1
 
 	VersionDefault = 0
 )
@@ -57,8 +57,7 @@ var (
 		"all":    IntervalAll,
 	}
 
-	ErrUndefinedSort       = errors.New("undefined sort")
-	ErrSearchQueryTooShort = errors.New("search query too short")
+	ErrUndefinedSort = errors.New("undefined sort")
 
 	// Ensure params types satisfy the interface
 	_ Param = &ListParams{}
@@ -67,7 +66,7 @@ var (
 )
 
 type Param interface {
-	ForValues(url.Values) error
+	ForValues(uint8, url.Values) error
 	CacheKey() []string
 }
 
@@ -79,12 +78,18 @@ func CacheKey(name string, val interface{}) string {
 // Global params
 //
 type ListParams struct {
-	Limit           int
-	Offset          int
+	Limit  int
+	Offset int
+
+	ID        *ids.ID
+	Query     string
+	StartTime time.Time
+	EndTime   time.Time
+
 	DisableCounting bool
 }
 
-func (p *ListParams) ForValues(q url.Values) (err error) {
+func (p *ListParams) ForValues(version uint8, q url.Values) (err error) {
 	p.Limit, err = GetQueryInt(q, KeyLimit, PaginationDefaultLimit)
 	if err != nil {
 		return err
@@ -93,24 +98,58 @@ func (p *ListParams) ForValues(q url.Values) (err error) {
 	if err != nil {
 		return err
 	}
-	p.DisableCounting, err = GetQueryBool(q, KeyDisableCount, false)
+
+	p.ID, err = GetQueryID(q, KeyID)
 	if err != nil {
 		return err
 	}
+
+	p.StartTime, err = GetQueryTime(q, KeyStartTime)
+	if err != nil {
+		return err
+	}
+	p.StartTime = p.StartTime.Round(TransactionRoundDuration)
+
+	p.EndTime, err = GetQueryTime(q, KeyEndTime)
+	if err != nil {
+		return err
+	}
+	if p.EndTime.IsZero() {
+		p.EndTime = time.Now().UTC()
+	}
+	p.EndTime = p.EndTime.Round(TransactionRoundDuration)
+
+	p.Query = GetQueryString(q, KeySearchQuery, "")
+
+	if p.DisableCounting, err = GetQueryBool(q, KeyDisableCount, false); err != nil {
+		return err
+	}
+
+	// Always disable counting for v2+
+	if version >= 2 {
+		p.DisableCounting = true
+	}
+
 	return nil
 }
 
 func (p *ListParams) CacheKey() []string {
-	return []string{
+	var keys []string
+	if p.ID != nil {
+		keys = append(keys, CacheKey(KeyID, p.ID.String()))
+	}
+
+	return append(keys,
 		CacheKey(KeyLimit, p.Limit),
 		CacheKey(KeyOffset, p.Offset),
 
-		// inject the DisableCount to the key..  Makes sure cache hits will return answer matching request
-		CacheKey(KeyDisableCount, p.DisableCounting),
-	}
+		CacheKey(KeyStartTime, p.StartTime.Unix()),
+		CacheKey(KeyEndTime, p.EndTime.Unix()),
+		CacheKey(KeySearchQuery, p.Query),
+		CacheKey(KeyDisableCount, p.DisableCounting))
 }
 
-func (p ListParams) Apply(b *dbr.SelectBuilder) *dbr.SelectBuilder {
+func (p ListParams) Apply(listTable string, b *dbr.SelectBuilder) *dbr.SelectBuilder {
 	if p.Limit > PaginationMaxLimit {
 		p.Limit = PaginationMaxLimit
 	}
@@ -120,5 +159,14 @@ func (p ListParams) Apply(b *dbr.SelectBuilder) *dbr.SelectBuilder {
 	if p.Offset != 0 {
 		b.Offset(uint64(p.Offset))
 	}
+
+	if p.ID != nil {
+		b.Where(listTable+".id = ?", p.ID.String()).Limit(1)
+	}
+
+	if p.Query != "" {
+		b.Where(dbr.Like(listTable+".id", p.Query+"%"))
+	}
+
 	return b
 }
